+++ src/node/node_state.h
@@ -16,10 +16,8 @@
 #include "ccf/pal/uvm_endorsements.h"
 #include "ccf/service/node_info_network.h"
 #include "ccf/service/reconfiguration_type.h"
-#include "ccf/service/tables/acme_certificates.h"
 #include "ccf/service/tables/service.h"
 #include "ccf/tx.h"
-#include "ccf_acme_client.h"
 #include "consensus/aft/raft.h"
 #include "consensus/ledger_enclave.h"
 #include "crypto/certs.h"
@@ -174,15 +172,6 @@ namespace ccf
     // Set to the snapshot seqno when a node starts from one and remembered for
     // the lifetime of the node
     ccf::kv::Version startup_seqno = 0;
-
-    // ACME certificate endorsement client
-    std::map<NodeInfoNetwork::RpcInterfaceID, std::shared_ptr<ACMEClient>>
-      acme_clients;
-    std::map<
-      NodeInfoNetwork::RpcInterfaceID,
-      std::shared_ptr<ACMEChallengeHandler>>
-      acme_challenge_handlers;
-    size_t num_acme_interfaces = 0;
 
     std::shared_ptr<ccf::kv::AbstractTxEncryptor> make_encryptor()
     {
@@ -603,8 +592,6 @@ namespace ccf
       setup_history();
       setup_snapshotter();
       setup_encryptor();
-
-      setup_acme_clients();
 
       initiate_quote_generation();
 
@@ -1522,80 +1509,6 @@ namespace ccf
         ccf::kv::CommittableTx::TxFlag::SNAPSHOT_AT_NEXT_SIGNATURE);
     }
 
-    void trigger_acme_refresh(
-      ccf::kv::Tx& tx,
-      const std::optional<std::vector<std::string>>& interfaces =
-        std::nullopt) override
-    {
-      if (!network.identity)
-      {
-        return;
-      }
-
-      num_acme_interfaces = 0;
-
-      for (const auto& [iname, interface] : config.network.rpc_interfaces)
-      {
-        if (
-          !interface.endorsement ||
-          interface.endorsement->authority != Authority::ACME ||
-          !interface.endorsement->acme_configuration)
-        {
-          continue;
-        }
-
-        num_acme_interfaces++;
-
-        if (
-          !interfaces ||
-          std::find(interfaces->begin(), interfaces->end(), iname) !=
-            interfaces->end())
-        {
-          auto challenge_frontend = find_acme_challenge_frontend();
-
-          const std::string& cfg_name =
-            *interface.endorsement->acme_configuration;
-          auto cit = config.network.acme->configurations.find(cfg_name);
-          if (cit == config.network.acme->configurations.end())
-          {
-            LOG_INFO_FMT("Unknown ACME configuration '{}'", cfg_name);
-            continue;
-          }
-
-          if (
-            !cit->second.directory_url.empty() &&
-            acme_clients.find(cfg_name) == acme_clients.end())
-          {
-            const auto& cfg = cit->second;
-
-            auto client = std::make_shared<ACMEClient>(
-              cfg_name,
-              cfg,
-              rpc_map,
-              rpcsessions,
-              challenge_frontend,
-              network.tables,
-              node_sign_kp);
-
-            auto chit = acme_challenge_handlers.find(iname);
-            if (chit != acme_challenge_handlers.end())
-            {
-              client->install_custom_challenge_handler(chit->second);
-            }
-
-            acme_clients.emplace(cfg_name, client);
-          }
-
-          auto client = acme_clients[cfg_name];
-          if (client && !client->has_active_orders())
-          {
-            client->get_certificate(
-              make_key_pair(network.identity->priv_key), true);
-          }
-        }
-      }
-    }
-
     void transition_service_to_open(
       ccf::kv::Tx& tx,
       AbstractGovernanceEffects::ServiceIdentities identities) override
@@ -2084,28 +1997,6 @@ namespace ccf
       return find_frontend(ActorsType::users)->is_open();
     }
 
-    std::shared_ptr<ACMERpcFrontend> find_acme_challenge_frontend()
-    {
-      auto acme_challenge_opt = rpc_map->find(ActorsType::acme_challenge);
-      if (!acme_challenge_opt)
-      {
-        throw std::runtime_error("Missing ACME challenge frontend");
-      }
-      return std::static_pointer_cast<ACMERpcFrontend>(*acme_challenge_opt);
-    }
-
-    void open_acme_challenge_frontend()
-    {
-      if (config.network.acme && !config.network.acme->configurations.empty())
-      {
-        auto fe = find_frontend(ActorsType::acme_challenge);
-        if (fe)
-        {
-          fe->open();
-        }
-      }
-    }
-
     std::vector<uint8_t> serialize_create_request(
       View create_view, bool create_consortium = true)
     {
@@ -2565,34 +2456,6 @@ namespace ccf
               LOG_INFO_FMT("Service open at seqno {}", hook_version);
             }
           }));
-
-      network.tables->set_global_hook(
-        network.acme_certificates.get_name(),
-        network.acme_certificates.wrap_commit_hook(
-          [this](
-            ccf::kv::Version hook_version, const ACMECertificates::Write& w) {
-            for (auto const& [interface_id, interface] :
-                 config.network.rpc_interfaces)
-            {
-              if (interface.endorsement->acme_configuration)
-              {
-                auto cit = w.find(*interface.endorsement->acme_configuration);
-                if (cit != w.end())
-                {
-                  LOG_INFO_FMT(
-                    "ACME: new certificate for interface '{}' with "
-                    "configuration '{}'",
-                    interface_id,
-                    *interface.endorsement->acme_configuration);
-                  rpcsessions->set_cert(
-                    Authority::ACME,
-                    *cit->second,
-                    network.identity->priv_key,
-                    cit->first);
-                }
-              }
-            }
-          }));
     }
 
     ccf::kv::Version get_last_recovered_signed_idx() override
@@ -2806,63 +2669,6 @@ namespace ccf
         ::consensus::ledger_truncate, to_host, idx, recovery_mode);
     }
 
-    void setup_acme_clients()
-    {
-      if (!config.network.acme || config.network.acme->configurations.empty())
-      {
-        return;
-      }
-
-      open_acme_challenge_frontend();
-
-      const auto& ifaces = config.network.rpc_interfaces;
-      num_acme_interfaces =
-        std::count_if(ifaces.begin(), ifaces.end(), [](const auto& id_iface) {
-          return id_iface.second.endorsement->authority == Authority::ACME;
-        });
-
-      if (num_acme_interfaces > 0)
-      {
-        using namespace threading;
-
-        // Start task to periodically check whether any of the certs are
-        // expired.
-        auto msg = std::make_unique<::threading::Tmsg<NodeStateMsg>>(
-          [](std::unique_ptr<::threading::Tmsg<NodeStateMsg>> msg) {
-            auto& state = msg->data.self;
-
-            if (state.consensus && state.consensus->can_replicate())
-            {
-              if (state.acme_clients.size() != state.num_acme_interfaces)
-              {
-                auto tx = state.network.tables->create_tx();
-                state.trigger_acme_refresh(tx);
-                tx.commit();
-              }
-              else
-              {
-                for (auto& [cfg_name, client] : state.acme_clients)
-                {
-                  if (client)
-                  {
-                    client->check_expiry(
-                      state.network.tables, state.network.identity);
-                  }
-                }
-              }
-            }
-
-            auto delay = std::chrono::minutes(1);
-            ::threading::ThreadMessaging::instance().add_task_after(
-              std::move(msg), delay);
-          },
-          *this);
-
-        ::threading::ThreadMessaging::instance().add_task_after(
-          std::move(msg), std::chrono::seconds(2));
-      }
-    }
-
     void seal_ledger_secret(const VersionedLedgerSecret& ledger_secret)
     {
       seal_ledger_secret(ledger_secret.first, ledger_secret.second);
@@ -2924,13 +2730,6 @@ namespace ccf
     virtual ccf::crypto::Pem get_network_cert() override
     {
       return network.identity->cert;
-    }
-
-    virtual void install_custom_acme_challenge_handler(
-      const ccf::NodeInfoNetwork::RpcInterfaceID& interface_id,
-      std::shared_ptr<ACMEChallengeHandler> h) override
-    {
-      acme_challenge_handlers[interface_id] = h;
     }
 
     // Stop-gap until it becomes easier to use other HTTP clients